Organized and annotated some updater code.
