+++ FOCollections/UITableView/FOTableViewController.swift
@@ -307,233 +307,3 @@ open class FOTableViewController: UIViewController, UITableViewDelegate {
     }
     
 }
-
-extension FOTableViewController {
-    
-    func animateUpdate(_ to: [FOTableSection], with animation: UITableViewRowAnimation = .automatic, duration: TimeInterval = 0.3, completion: (() -> ())? = nil) {
-        queue.addOperation(FOCompletionOperation(work: {
-            [weak self] (operation) -> Void in
-            UIView.animate(withDuration: duration, animations: {
-                if let this = self {
-                    //let date = Date()
-                    var doneCount = 0
-                    
-                    func processDone() {
-                        doneCount += 1
-                        
-                        if doneCount == 2 {
-                            operation.finish()
-                            completion?()
-                        }
-                    }
-                    
-                    func updateSections() {
-                        CATransaction.begin()
-                        CATransaction.setCompletionBlock {
-                            //print("first", Date().timeIntervalSinceReferenceDate - date.timeIntervalSinceReferenceDate)
-                            processDone()
-                        }
-                        
-                        this.tableView.beginUpdates()
-                        
-                        let fromSections0 = this.dataSource.sections
-                        let fromPaths0 = this.dataSource.statePaths()
-                        let toPaths0 = this.dataSource.statePaths(sections: to)
-                        let updater0 = FOCollectionUpdater(from: fromPaths0, to: toPaths0)
-                        let update0 = updater0.update(index: 0)
-                        
-                        this.updateSections(update: update0, with: animation)
-                        
-                        let transformed = this.transform(fromSections: fromSections0, toSections: to, update: update0)
-                        
-                        _ = this.dataSource.clearAllItems(this.tableView)
-                        this.dataSource.insertSections(transformed, atIndexes: IndexSet(integersIn: 0..<transformed.count), tableView: this.tableView, viewController: this)
-                        
-                        this.tableView.endUpdates()
-                        
-                        CATransaction.commit()
-                    }
-                    
-                    func updateItems() {
-                        CATransaction.begin()
-                        CATransaction.setCompletionBlock {
-                            //print("second", Date().timeIntervalSinceReferenceDate - date.timeIntervalSinceReferenceDate)
-                            processDone()
-                        }
-                        
-                        this.tableView.beginUpdates()
-                        
-                        let fromPaths1 = this.dataSource.statePaths()
-                        let toPaths1 = this.dataSource.statePaths(sections: to)
-                        let updater1 = FOCollectionUpdater(from: fromPaths1, to: toPaths1)
-                        let update1 = updater1.update(index: 1)
-                        
-                        this.updateItems(update: update1, with: animation)
-                        
-                        _ = this.dataSource.clearAllItems(this.tableView)
-                        this.dataSource.insertSections(to, atIndexes: IndexSet(integersIn: 0..<to.count), tableView: this.tableView, viewController: this)
-                        
-                        this.tableView.endUpdates()
-                        
-                        CATransaction.commit()
-                    }
-                    
-                    updateSections()
-                    updateItems()
-                    this.refreshVisibleCells()
-                }
-            })
-        }, queue: DispatchQueue.main))
-    }
-    
-    fileprivate func transform(fromSections: [FOTableSection], toSections: [FOTableSection], update: Update) -> [FOTableSection] {
-        var result = fromSections
-        let iIdentifiers: [String] = update.insertions != nil ? itemIdentifiers(fromSections) : [String]()
-        
-        update.deletions?.forEach({
-            path in
-            if let index = index(path: path, in: result) {
-                result.remove(at: index)
-            }
-        })
-        
-        update.insertions?.forEach({
-            path in
-            if let index = index(path: path, in: toSections) {
-                let section = removeDuplicateItems(toSections[index], itemIdentifiers: iIdentifiers)
-                result.insert(section, at: index)
-            }
-        })
-        
-        toSections.enumerated().forEach {
-            (s) in
-            if
-                let index = result.index(of: s.element),
-                index != s.offset
-            {
-                result.move(at: index, to: s.offset)
-            }
-        }
-        
-        return result
-    }
-    
-    private func itemIdentifiers(_ sections: [FOTableSection]) -> [String] {
-        return sections.reduce([String](), { (sResult, section) -> [String] in
-            if let items = section.items {
-                return sResult + items.reduce([String](), { (iResult, item) -> [String] in
-                    if let identifier = item.identifier {
-                        return iResult + [identifier]
-                    } else {
-                        return iResult
-                    }
-                })
-            } else {
-                return sResult
-            }
-        })
-    }
-    
-    private func itemIdentifiers(_ section: FOTableSection) -> [String] {
-        if let items = section.items {
-            return items.reduce([String](), { (iResult, item) -> [String] in
-                if let identifier = item.identifier {
-                    return iResult + [identifier]
-                } else {
-                    return iResult
-                }
-            })
-        } else {
-            return [String]()
-        }
-    }
-    
-    private func removeDuplicateItems(_ section: FOTableSection, itemIdentifiers: [String]) -> FOTableSection {
-        let sIdentifiers = self.itemIdentifiers(section)
-        let iIdentifiers = Array(itemIdentifiers)
-        let duplicates = iIdentifiers.filter{sIdentifiers.contains($0)}
-        var result = section
-        
-        if
-            !duplicates.isEmpty,
-            let section = section.copy() as? FOTableSection
-        {
-            result = section
-            result.items = result.items?.filter({
-                item in
-                if let identifier = item.identifier {
-                    return !duplicates.contains(identifier)
-                } else {
-                    return true
-                }
-            })
-        }
-        
-        return result
-    }
-    
-    fileprivate func index(path: StatePath, in sections: [FOTableSection]) -> Int? {
-        return sections.index(where: {
-            (section) -> Bool in
-            section.identifier == path.identifierPath.identifiers[0]
-        })
-    }
-    
-    fileprivate func updateSections(update: Update, with animation: UITableViewRowAnimation) {
-        if
-            let deletions = update.deletions,
-            deletions.count > 0
-        {
-            tableView.deleteSections(IndexSet(deletions.map{$0.indexPath[0]}), with: animation)
-        }
-        
-        if
-            let insertions = update.insertions,
-            insertions.count > 0
-        {
-            tableView.insertSections(IndexSet(insertions.map{$0.indexPath[0]}), with: animation)
-        }
-        
-        if
-            let moves = update.moves,
-            moves.count > 0
-        {
-            for move in moves {
-                tableView.moveSection(move.from.indexPath[0], toSection: move.to.indexPath[0])
-            }
-        }
-    }
-    
-    fileprivate func updateItems(update: Update, with animation: UITableViewRowAnimation) {
-        if
-            let deletions = update.deletions,
-            deletions.count > 0
-        {
-            tableView.deleteRows(at: deletions.map{$0.indexPath}, with: animation)
-        }
-        
-        if
-            let insertions = update.insertions,
-            insertions.count > 0
-        {
-            tableView.insertRows(at: insertions.map{$0.indexPath}, with: animation)
-        }
-        
-        if
-            let moves = update.moves,
-            moves.count > 0
-        {
-            for move in moves {
-                tableView.moveRow(at: move.from.indexPath, to: move.to.indexPath)
-            }
-        }
-    }
-    
-}
-
-extension Array {
-    mutating func move(at oldIndex: Int, to newIndex: Int) {
-        insert(remove(at: oldIndex), at: newIndex)
-    }
-}
-
